msm: kgsl: Fix overflow issue by checking user supplied count
User supplied count can cause signed integer overflow when passed to
dma_fence_array_create. So make sure that count is less than INT_MAX.

+++ drivers/gpu/msm/kgsl_timeline.h
@@ -108,6 +108,6 @@ static inline void kgsl_timeline_put(struct kgsl_timeline *timeline)
  * encapsulated timeline fences to expire.
  */
 struct dma_fence *kgsl_timelines_to_fence_array(struct kgsl_device *device,
-		u64 timelines, u64 count, u64 usize, bool any);
+		u64 timelines, u32 count, u64 usize, bool any);
 
 #endif
